requirements.txt
torch==2.6.0
torchvision==0.21.0
einops==0.8.1
liger-kernel
transformers==4.49.0
qwen_vl_utils==0.0.10
safetensors==0.4.5
diffusers[torch]==0.32.2
pillow==11.1.0
numpy==1.25.0
opencv-python==4.8.1.78
omegaconf==2.3.0
matplotlib==3.10.1
tabulate==0.9.0
tqdm==4.67.1
rich==13.9.4
